Squash PR #8201: Remove Qt dependencies from OpenMS core (rebased on develop)
Original PR: Replace Qt networking with cpp-httplib and eliminate Qt
dependencies from OpenMS core.

Networking changes (cpp-httplib) are excluded as develop already uses
libcurl (commit 3efb237112). The remaining changes remove Qt from:

- Date/DateTime: pure C++ replacing QDate/QDateTime
- String/DataValue/StringUtils: remove QString conversions
- ExternalProcess: Boost.Process (Windows) / fork-exec (POSIX)
- File: std::filesystem replacing QDir/QFileInfo
- SHA: new portable SHA-1 implementation
- FileWatcher: stub in core, full impl in GUI
- RWrapper: uses new ExternalProcess API
- TOPP tools: updated for std::string/std::vector APIs
- GUI: updated for changed core APIs

+++ src/openms/include/OpenMS/ANALYSIS/TOPDOWN/PeakGroup.h
@@ -258,7 +258,7 @@ namespace OpenMS
     int getMinNegativeIsotopeIndex() const;
 
     /// set index of this peak group
-    void setIndex(uint i);
+    void setIndex(size_t i);
 
     /**
      * @brief Set the two-dimensional quality score for this peak group.
